Modulized sortProducts functionality
