+++ products.html
@@ -172,6 +172,6 @@
     </div>
     <script src="js/bootstrap.bundle.min.js"></script>
     <script defer src="js/init.js"></script>
-    <script defer src="js/products.js"></script>
+    <script defer type="module" src="js/products.js"></script>
   </body>
 </html>
